functional pipeline without optimization

+++ calibration.py
@@ -3,36 +3,47 @@ import cv2
 import glob
 import pickle
 import matplotlib.pyplot as plt
+import os.path
 
-# prepare object points
-objp = np.zeros((9*6,3), np.float32)
-objp[:,:2] = np.mgrid[0:9, 0:6].T.reshape(-1,2)
+def calibrate():
+	pickle_path = "dist_pickle.p"
+	# if calibration already run, just unpickle the stored values
+	if os.path.exists(pickle_path):
+		dist_pickle = pickle.load(open(pickle_path, "rb"))
+		mtx = dist_pickle["mtx"]
+		dist = dist_pickle["dist"]
+	else:
+		# prepare object points
+		objp = np.zeros((9*6,3), np.float32)
+		objp[:,:2] = np.mgrid[0:9, 0:6].T.reshape(-1,2)
 
-objpoints = [] # 3d points in real world space
-imgpoints = [] # 2d points in image plane
+		objpoints = [] # 3d points in real world space
+		imgpoints = [] # 2d points in image plane
 
-# List of calibration images
-images = glob.glob('camera_cal/calibration*.jpg')
+		# List of calibration images
+		images = glob.glob('camera_cal/calibration*.jpg')
 
-for idx, fname in enumerate(images):
-    img = cv2.imread(fname)
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-    
-    # Find the chessboard corners
-    ret, corners = cv2.findChessboardCorners(gray, (9,6), None)
-    
-    # If found, add object points, image points
-    if ret == True:
-        objpoints.append(objp)
-        imgpoints.append(corners)
+		for idx, fname in enumerate(images):
+		    img = cv2.imread(fname)
+		    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+		    
+		    # Find the chessboard corners
+		    ret, corners = cv2.findChessboardCorners(gray, (9,6), None)
+		    
+		    # If found, add object points, image points
+		    if ret == True:
+		        objpoints.append(objp)
+		        imgpoints.append(corners)
 
 
-# Do camera calibration given object points and image points
-cal_img = cv2.imread(images[0])
-img_size = (img.shape[1], img.shape[0])
-ret, mtx, dist, rvecs, tvecs = cv2.calibrateCamera(objpoints, imgpoints, img_size, None, None)
+		# Do camera calibration given object points and image points
+		cal_img = cv2.imread(images[0])
+		img_size = (img.shape[1], img.shape[0])
+		ret, mtx, dist, rvecs, tvecs = cv2.calibrateCamera(objpoints, imgpoints, img_size, None, None)
 
-dist_pickle = {}
-dist_pickle["mtx"] = mtx
-dist_pickle["dist"] = dist
-pickle.dump( dist_pickle, open( "dist_pickle.p", "wb" ))
+		dist_pickle = {}
+		dist_pickle["mtx"] = mtx
+		dist_pickle["dist"] = dist
+		pickle.dump( dist_pickle, open( "dist_pickle.p", "wb" ))
+
+	return mtx, dist
